Move functionality to the right module and use keyword arguments.
